add run_app.py to start streamlit
