+++ run_app.sh
@@ -1,0 +1,1 @@
+streamlit run app.py --server.port 8080
